requirements/production.txt
django-filter==0.8
Django==1.7
djangorestframework==2.4.3
feedparser==5.1.3
jsonfield==1.0.0
nodeenv==0.11.1
